+++ src/screens/index.ts
@@ -1,6 +1,8 @@
 /* Modified by Lux Music: derived from the upstream LX Music Mobile source file. This file remains under Apache-2.0. See LICENSE-NOTICE.md. */
 
 export { default as Launch } from './Launch'
+export { default as Login } from './Login'
+export { default as Agreement } from './Agreement'
 export { default as Home } from './Home'
 export { default as PlayDetail } from './PlayDetail'
 export { default as SonglistDetail } from './SonglistDetail'