+++ src/types/app.d.ts
@@ -56,6 +56,8 @@ interface GlobalData {
   settingActiveId: SettingScreenIds
   isShowingLaunchScreen: boolean
 
+  _onLoginConfirmed?: () => void | Promise<void>
+
   /**
    * 首页是否正在滚动中，用于防止意外误触播放歌曲
    */